chore(settings): refactor for dynamic settings cards
